pyproject.toml
[project]
name = "riset-payment-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
]
